requirements.txt
mcp>=1.2.0
mysql-connector-python>=9.0.0
uvicorn>=0.30.0
